Changed to Differentiate between Doctor and Patient
Doctor now requires ID when logging in while the patient does not. The code differentiates between the different types of login based on the username. After comparing the username with the type of user they are it confirms whether an ID is required or not and requests for one if needed.