Edits to Server, Login and Signup
Changed some logic and fixed some errors in SignupAndErrorSyste.js. Will add comments soon. Additionally, made this a module and imported the necessary code to signup.js. Lastly, added a clearAllErrors() helper function to clear the red errors on successful signin. That way the password field isn't weird.

Changed some logic in the login.html file that was made into a separate file named login.js. This was causing some errors on login, hence not giving any feedback that the login was correct. This is all for patient login, doctor login has only been edited to be inline with the new format. Still needs to be corrected. Some other small errors were corrected.

LoginSystem.js now has correct logic for determining type. This is then also corrected in the authentication. Lastly, the saved format was changed to an array for ease of reading for both us and a computer. thus the logic was changed. Some console messages added to check if things are working correctly. Small errors corrected.

server.js now correctly posts type = 'patient' for patients that are signing up. This allows the login to work correctly. Doctor login format is different and might still need to be corrected later, it was not touched much. That logic can be found in LoginSystem.js.

Signup.html small errors changed. Your changes should remain.

signup.js just changed to be .json instead of .text.