Set Up Server to Work for Anyone from Repository
Server is now set up to work from command bar for anyone so long as file locations are not changed. I can update this if need be later on.

USAGE: npm start

This will start the server you should see something along the lines of the following if it works:
> cmsc-355-project@1.0.0 start
> node Sprint_1/server.js

Server is running on http://localhost:5500

It looks like everything has been changed but thats just due to how the express and body-parser are installed.

This update should make accessing an actual live server much easier for everyone and doesn't require a long explanation anymore.